Turn off csrf for open id failure
When conversation with openid does not go well user was redirected into
`failure` devise controller method. Unfortunately plgrid open id is not
able to redirect csrf token and this method requires it. It was solved
by turned off authenticity token verify for `failure` method. We can do
this since it does not do any modification on the server side, it is
only redirecting user into login page.